split products and product components
