+++ src/components/header.tsx
@@ -4,6 +4,7 @@ import GithubCorner from 'react-github-corner';
 import { Popover, Select } from 'antd';
 import classNames from 'classnames';
 import Search from './search';
+import Products from './products';
 import { docs } from '../../.antvisrc';
 import styles from './header.module.less';
 
@@ -14,225 +15,6 @@ interface HeaderProps {
   location?: Location;
   currentLangKey?: string;
 }
-
-interface ProductProps {
-  name?: string;
-  icon?: string;
-  description?: string;
-  url?: string;
-  links?: Array<{
-    name: string;
-    link: string;
-  }>;
-}
-
-const Product: React.FC<ProductProps> = ({
-  name,
-  icon,
-  url,
-  description,
-  links = [],
-}) => (
-  <li className={styles.product}>
-    <img alt={name} src={icon} />
-    <div className={styles.productContent}>
-      <a href={url} target="_blank">
-        <h4>{name}</h4>
-      </a>
-      <div className={styles.productDescription}>{description}</div>
-      <div className={styles.productLinks}>
-        {links.map(({ link, name }, i) => (
-          <a target="_blank" href={link} key={i}>
-            {name}
-          </a>
-        ))}
-      </div>
-    </div>
-  </li>
-);
-
-const Products: React.FC<any> = () => (
-  <div className={styles.products}>
-    <h3>基础产品</h3>
-    <ul>
-      <Product
-        name="G2"
-        url="/g2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-    </ul>
-    <h3>扩展产品</h3>
-    <ul>
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-      <Product
-        name="G2"
-        description="一套基于可视化编码的图形语法"
-        icon="https://gw.alipayobjects.com/zos/antfincdn/mpz%24sFOwuD/1852987f-0412-44e5-9d0d-75b06f392778.png"
-        links={[
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-          {
-            name: '关于',
-            link: '',
-          },
-        ]}
-      />
-    </ul>
-  </div>
-);
 
 const Header: React.FC<HeaderProps> = ({
   siteTitle = '',
